backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.7
bcrypt==4.1.2
alembic==1.13.1
pytest==7.4.4
httpx==0.26.0
requests==2.31.0
websockets==12.0

# 开发工具依赖
rich==13.7.0
sqlalchemy-schemadisplay==1.3
pydot==1.4.2
graphviz==0.20.1
eralchemy2==1.3.6 